provider: Pass list of CustomizeDiffFuncs to wrapper instead of special-casing transparent tagging.
