+++ internal/provider/tags_interceptor.go
@@ -5,8 +5,11 @@ package provider
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-provider-aws/internal/conns"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs/sdkdiag"
 	"github.com/hashicorp/terraform-provider-aws/internal/provider/interceptors"
 	tftags "github.com/hashicorp/terraform-provider-aws/internal/tags"
@@ -258,4 +261,58 @@ func (r tagsInterceptor) getIdentifier(d schemaResourceData) string {
 	}
 
 	return identifier
+}
+
+// setTagsAll is a CustomizeDiff function that calculates the new value for the `tags_all` attribute.
+func setTagsAll(ctx context.Context, d *schema.ResourceDiff, meta interface{}) error {
+	c := meta.(*conns.AWSClient)
+
+	if !d.GetRawPlan().GetAttr(names.AttrTags).IsWhollyKnown() {
+		if err := d.SetNewComputed(names.AttrTagsAll); err != nil {
+			return fmt.Errorf("setting tags_all to Computed: %w", err)
+		}
+		return nil
+	}
+
+	newTags := tftags.New(ctx, d.Get(names.AttrTags).(map[string]interface{}))
+	allTags := c.DefaultTagsConfig(ctx).MergeTags(newTags).IgnoreConfig(c.IgnoreTagsConfig(ctx))
+	if d.HasChange(names.AttrTags) {
+		if newTags.HasZeroValue() {
+			if err := d.SetNewComputed(names.AttrTagsAll); err != nil {
+				return fmt.Errorf("setting tags_all to Computed: %w", err)
+			}
+		}
+
+		if len(allTags) > 0 && (!newTags.HasZeroValue() || !allTags.HasZeroValue()) {
+			if err := d.SetNew(names.AttrTagsAll, allTags.Map()); err != nil {
+				return fmt.Errorf("setting new tags_all diff: %w", err)
+			}
+		}
+
+		if len(allTags) == 0 {
+			if err := d.SetNew(names.AttrTagsAll, allTags.Map()); err != nil {
+				return fmt.Errorf("setting new tags_all diff: %w", err)
+			}
+		}
+	} else {
+		if len(allTags) > 0 && !allTags.HasZeroValue() {
+			if err := d.SetNew(names.AttrTagsAll, allTags.Map()); err != nil {
+				return fmt.Errorf("setting new tags_all diff: %w", err)
+			}
+			return nil
+		}
+
+		var newTagsAll tftags.KeyValueTags
+		if v, ok := d.Get(names.AttrTagsAll).(map[string]interface{}); ok {
+			newTagsAll = tftags.New(ctx, v)
+		}
+		if len(allTags) > 0 && !newTagsAll.DeepEqual(allTags) && allTags.HasZeroValue() {
+			if err := d.SetNewComputed(names.AttrTagsAll); err != nil {
+				return fmt.Errorf("setting tags_all to Computed: %w", err)
+			}
+			return nil
+		}
+	}
+
+	return nil
 }
